Alot of change in this one.
Found a meta button that was blocked in the package, which contained the script to make it interactable with hand interaction. Working on the UI Builder for the game object list to spawn in the scene. Added a sec. Build a file for hiding the effect mesh, and hide the mesh function so it looks more MR Immersive.