Added Scene From the Meta SDK Package
Found some scene in the package that can't be open while living in the package . To fix that we just copy and paste those into the asset file. That how to fix the read only thing of the project package. Also fix the teddy bear where it is grabbable it was the Pointable Element that keep it from grabbed from the interactable hand.